requirements.txt
psycopg2-binary==2.9.3
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
python-jose==3.3.0
sqlalchemy==1.4.36
elasticsearch==7.14.0
elasticsearch-dsl==7.4.0
numpy==1.22.3
cmake==3.22.4
opencv-python==4.5.5.64
face_recognition==1.3.0
aiofiles==0.8.0
CouchDB==1.2
psycopg2-binary==2.9.3
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
python-jose==3.3.0
sqlalchemy==1.4.35
elasticsearch==7.14.0
elasticsearch-dsl==7.4.0
numpy==1.22.3
cmake==3.22.3
opencv-python==4.5.5.64
aiofiles==0.8.0
tensorflow==2.8.0
pandas==1.4.2
protobuf==3.20.1
fer==22.4.0
CouchDB==1.2